Got rid of types/calendar.ts and /utils/calendarUtils.ts because of server side issues. Calendar is now using real risk endpoint and the gemini service for study plan generation.
