+++ backend/app/api/v1/api.py
@@ -16,7 +16,7 @@ api_router.include_router(report.router, prefix="", tags=["report"])
 api_router.include_router(settings.router, prefix="/settings", tags=["settings"])
 api_router.include_router(about.router, prefix="/about", tags=["about"])
 api_router.include_router(question.router, prefix="/question", tags=["question"])
-#api_router.include_router(calendar.router, prefix="/calendar", tags=["calendar"])
+api_router.include_router(calendar.router, prefix="/calendar", tags=["calendar"])
 api_router.include_router(about.router, prefix="/about", tags=["about"])
 api_router.include_router(rag.router, prefix="/rag", tags=["rag"])
 api_router.include_router(gemini.router)
